ref(scheduler) move module
