+++ backend/api_server/src/scheduler/tasks/mod.rs
@@ -1,4 +1,3 @@
-use crate::blog::scheduler::task_func::TaskFunc;
 use chrono::{DateTime, Utc};
 
 pub mod basic;
